rename `sigma_proof::ctxt_comm_equality_proof` to non-abbreviated name
