+++ zk-token-sdk/src/sigma_proofs/ciphertext_commitment_equality_proof.rs
